+++ CHANGELOG.md
@@ -3,9 +3,10 @@ STRNaming Changelog
 
 ### Version 1.1.2 (TBD)
 * Updated CE allele numbering of DYS612 (+6 units).
+* Added reference structure for SE33.
 * Updated hardcoded reference length adjustment table to suppress second
   structure 5' of DYS522.
-* Added reference structure for SE33.
+* Added double-click-to-toggle-text-alignment feature to HTML output.
 
 ### Version 1.1.1 (19 July 2021)
 * Fixed an issue with CE allele numbering that occurred for reporting ranges